# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ArtFi API
  version: 0.4.0
  description: Stage 4 Base Sepolia preview contract. Wallet intents never hold private keys; authenticated indexer writes are durable and replay-safe.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/governance/proposals:
    get:
      operationId: getGovernanceProposals
      description: Durable chain-indexed governance proposals. No fixture fallback is permitted.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Indexed governance proposal records
          content:
            application/json:
              schema:
                type: object
                required: [data, membershipAuthority, schemaVersion]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/GovernanceProposal" }
                  membershipAuthority:
                    { type: string, const: onchain-rwa-token-snapshot }
                  schemaVersion: { type: string, const: "1" }
        "503": { $ref: "#/components/responses/Problem" }
...
components:
...
  responses:
    Problem:
      description: RFC 9457-style request failure
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    GovernanceProposal:
      type: object
      required:
        [
          actionTarget,
          approvalNumerator,
          approvalDenominator,
          governorAddress,
          proposalId,
          proposalKind,
          proposerAddress,
          rwaVerified,
          status,
          voteStart,
          voteEnd,
        ]
      properties:
        actionTarget: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        approvalNumerator: { type: integer, enum: [500000, 666667, 800000] }
        approvalDenominator: { type: integer, const: 1000000 }
        evidenceHash: { type: string, pattern: "^0x[0-9a-fA-F]{64}$" }
        evidenceUri: { type: string, format: uri }
        governorAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        proposalId: { type: string }
        proposalKind:
          {
            type: string,
            enum: [market_migration, physical_action, forced_buyout],
          }
        proposerAddress: { type: string, pattern: "^0x[0-9a-fA-F]{40}$" }
        rwaVerified: { type: boolean }
        status:
          {
            type: string,
            enum:
              [
                pending,
                active,
                defeated,
                succeeded,
                queued,
                executed,
                cancelled,
              ],
          }
        voteStart: { type: integer, minimum: 0 }
        voteEnd: { type: integer, minimum: 0 }
...
    Problem:
      type: object
      required: [type, title, status, detail, requestId]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        requestId: { type: string }